Renamed TooltipComponent to Tooltip
