+++ kafka-ui-react-app/src/components/common/Tooltip/Tooltip.styled.ts
