+++ kafka-ui-react-app/src/components/common/Tooltip/Tooltip.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import Tippy from '@tippyjs/react';
 
-import * as S from './TooltipComponent.styled';
+import * as S from './Tooltip.styled';
 
 export interface PropsTypes {
   value: string;
   messageTooltip: string;
 }
-const TooltipComponent: React.FC<PropsTypes> = ({ value, messageTooltip }) => {
+const Tooltip: React.FC<PropsTypes> = ({ value, messageTooltip }) => {
   return (
     <Tippy content={<S.MessageTooltip>{messageTooltip}</S.MessageTooltip>}>
       <span>{value}</span>
     </Tippy>
   );
 };
-export default TooltipComponent;
+export default Tooltip;
